requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.64.0
tensorboard>=2.10.0
Pillow>=9.0.0
scipy>=1.9.0
scikit-learn>=1.1.0
